Statically link dependencies
Use git submodules to include our main dependencies (Catch2 and entt) in
the source code directory. Also fetch Raylib with FetchContent since we
want to get rid of Raylib eventually.

This will make it not an absolute nightmare to work on this project
between computers, and in theory will allow us to run Github Actions
workflows on this project.