Add Velocity Component
Create a new component called "Velocity" to generalize things that move
in a specific velocity every frame, and create a system that applies
that velocity to an entity's position in an isolated, testable, and
framerate-independant way.

Also remove the speed/velocity calculations from the UserProjectile
component so we can use the new Velocity component instead.