Use DetectCollision
Create an enemy type that we can shoot at and test the `DetectCollision`
system with. This commit also includes various refinements to the
messaging and collision systems. There were a few messaging bugs that
slipped through in particular:
- Pooled messages were not freed when their parent inboxes were
  unregistered
- The MessageInbox component was not pointer stable and so the
  MessageDispatchService was not sending events to the proper inboxes

Fix these and also improve how we iterate over MessageCollections by
providing a custom iterator to use.