Add AudioService to GameServices/RenderingServices
Also fix some memory allocation errors in the SoloudAudioDevice. The
`SoLoud::Soloud` instance needs to be freed after all audio sources and
in the previous implementation it could create errors on shutdown.

This involved me debugging a "stack smashing" error for a while until I
realized that I was linking a locally installed version of SoLoud
instead of the static version provided in Git submodules. Oopsy.